+++ plugin/AndroidRemote/Entities/RadioStation.cs
@@ -1,0 +1,14 @@
+﻿using System.Runtime.Serialization;
+
+namespace MusicBeePlugin.AndroidRemote.Entities
+{
+    [DataContract]
+    public class RadioStation
+    {
+        [DataMember(Name = "name")]
+        public string Name { get; set; }
+
+        [DataMember(Name = "url")]
+        public string Url { get; set; }
+    }
+}